requirements.txt
pymodbus==2.5.3
paho-mqtt==1.6.1
